+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### x.y.z
+- :star: Add Pushover trigger
+
 ### 3.1.3
 - :fire: Fix IFTTT trigger & documentation